+++ parser_library/src/diagnostic_op.h
@@ -329,6 +329,10 @@ struct diagnostic_op
 
     static diagnostic_op error_A173_invalid_psect(const range& range);
 
+    static diagnostic_op error_A174_EQU_prog_type_val_format(const range& range);
+
+    static diagnostic_op error_A175_data_prog_type_deps(const range& range);
+
     // operand parameters
 
     static diagnostic_op error_A200_SCOPE_param(std::string_view instr_name, const range& range);