+++ parser_library/src/expressions/data_definition.h
@@ -79,6 +79,8 @@ struct data_definition final : public context::dependable
     int32_t get_scale_attribute(context::dependency_solver& info, diagnostic_op_consumer& diags) const;
     uint32_t get_length_attribute(context::dependency_solver& info, diagnostic_op_consumer& diags) const;
     int32_t get_integer_attribute(context::dependency_solver& info, diagnostic_op_consumer& diags) const;
+    context::symbol_attributes::program_type get_program_attribute(
+        context::dependency_solver& info, diagnostic_op_consumer& diags) const;
     context::symbol_attributes get_symbol_attributes(
         context::dependency_solver& info, diagnostic_op_consumer& diags) const;